// src/app/models/Visit.ts
export class Visit {
  doctorName: string;
  visitDate: Date;
  addODSphere: any;
  addODVA: string;
  addOSIPD: string;
  addOSSphere: string;
  addOSVA: string;
  newGlassIPD: string;
  newGlassODAxis: string;
  newGlassODCylinder: string;
  newGlassODSphere: string;
  newGlassODVA: string;
  newGlassOSAxis: string;
  newGlassOSCylinder: string;
  newGlassOSSphere: string;
  newGlassOSVA: string;
  patientName: string;
}